% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{calcul_distribution_semaine}
\alias{calcul_distribution_semaine}
\title{Calculer la distribution des trajets par jour de la semaine}
\usage{
calcul_distribution_semaine(trajet, filtre = TRUE)
}
\arguments{
\item{trajet}{Un data.frame respectant le schéma de df_velo.}

\item{filtre}{Booléen, TRUE par défaut.}
}
\value{
Un data.frame avec la distribution.
}
\description{
Calculer la distribution des trajets par jour de la semaine
}
