% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_boucle}
\alias{compter_nombre_boucle}
\title{Compter le nombre de boucles distinctes}
\usage{
compter_nombre_boucle(trajet)
}
\arguments{
\item{trajet}{Un data.frame respectant le schéma de df_velo.}
}
\value{
Un entier représentant le nombre de boucles distinctes.
}
\description{
Compte le nombre de boucles de comptage uniques dans le jeu de données.
}
