% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_trajets}
\alias{compter_nombre_trajets}
\title{Compter le nombre total de trajets}
\usage{
compter_nombre_trajets(trajet)
}
\arguments{
\item{trajet}{Un data.frame respectant le schéma de df_velo.}
}
\value{
Un nombre représentant le total des trajets.
}
\description{
Calcule la somme de la colonne Total.
}
