% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtre_anomalie}
\alias{filtre_anomalie}
\title{Filtrer les anomalies dans les données vélo}
\usage{
filtre_anomalie(trajet)
}
\arguments{
\item{trajet}{Un data.frame respectant le schéma de df_velo.}
}
\value{
Un data.frame nettoyé.
}
\description{
Supprime les lignes contenant des anomalies et les valeurs aberrantes.
}
