% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtrer_trajet}
\alias{filtrer_trajet}
\title{Filtrer les trajets par numéro de boucle}
\usage{
filtrer_trajet(trajet, boucle)
}
\arguments{
\item{trajet}{Un data.frame respectant le schéma de df_velo.}

\item{boucle}{Un vecteur de numéros de boucle à conserver.}
}
\value{
Un data.frame filtré contenant uniquement les boucles demandées.
}
\description{
Filtre un data.frame pour ne garder que les boucles sélectionnées.
}
\examples{
filtrer_trajet(trajet = df_velo, boucle = c("880", "881"))

}
