% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{trouver_trajet_max}
\alias{trouver_trajet_max}
\title{Trouver le trajet avec le maximum de passages}
\usage{
trouver_trajet_max(trajet)
}
\arguments{
\item{trajet}{Un data.frame respectant le schéma de df_velo.}
}
\value{
Un data.frame avec les infos du trajet max.
}
\description{
Trouver le trajet avec le maximum de passages
}
