% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{calcul_distribution_semaine}
\alias{calcul_distribution_semaine}
\title{Calculer la distribution des trajets par jour de la semaine}
\usage{
calcul_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{Un data.frame respectant le schéma de df_velo.}
}
\value{
Un data.frame avec le jour de la semaine et le nombre de trajets.
}
\description{
Compte la somme des trajets pour chaque jour de la semaine.
}
