% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtre_anomalie}
\alias{filtre_anomalie}
\title{Filtrer les anomalies dans les données vélo}
\usage{
filtre_anomalie(trajet)
}
\arguments{
\item{trajet}{Un data.frame respectant le schéma de df_velo.}
}
\value{
Un data.frame nettoyé sans les anomalies.
}
\description{
Supprime les lignes contenant des anomalies (Forte ou Faible),
ainsi que les valeurs aberrantes (Total > 10000 ou Total <= 0).
}
