% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{plot_distribution_semaine}
\alias{plot_distribution_semaine}
\title{Afficher la distribution des trajets par jour de la semaine}
\usage{
plot_distribution_semaine(trajet, filtre = TRUE)
}
\arguments{
\item{trajet}{Un data.frame respectant le schéma de df_velo.}

\item{filtre}{Booléen, TRUE par défaut.}
}
\value{
Un objet ggplot.
}
\description{
Afficher la distribution des trajets par jour de la semaine
}
