% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{trouver_trajet_max}
\alias{trouver_trajet_max}
\title{Trouver le trajet avec le maximum de passages}
\usage{
trouver_trajet_max(trajet)
}
\arguments{
\item{trajet}{Un data.frame respectant le schéma de df_velo.}
}
\value{
Un data.frame avec le nom de la boucle, le jour, le total,
la moyenne par jour et la moyenne par boucle.
}
\description{
Identifie la paire boucle-jour ayant le plus grand nombre de passages,
après filtrage des anomalies. Retourne aussi la moyenne par jour
et la moyenne par boucle.
}
